Magento 2.4.6 support (#5)
In short the aim is to have the module support Magento version 2.4.6. 

Major changes related to this:

- Magento\Framework\HTTP\ZendClient has been dropped in 2.4.6, which has
been used by this module. To still support 2.4.4 and 2.4.5 which still
have the class available the usage has been refactored to support all of
these three versions. Details about this can be checked from
Model/CustobarApi/README.md
- Support for Magento versions older than 2.4.4 has been dropped as
2.4.4 is the oldest officially supported version by Adobe. This means we
could get rid of the deprecated InstallSchema and UpgradeSchema usages
and replace these with the db_schema.xml definitions. This is
technically a breaking change according to semantic versioning, so this
is the main reason I updated the version in the README to be 3.0.0.

Minor changes added alongside this:

- Brought test definitions up to date so that they can be run in the
currently supported Magento test frameworks, for some reason this has
been left out from earlier Magento version support changes
- Fixed several code sniffer related warnings to make sure that catching
actual errors is easier. For example, lack of comments, non scoped
constants etc.
- Fixed issue with order placement time not exporting with correct value
to Custobar API
- Fixed issue with sku resolving not always working Commerce edition due
to the row_id and entity_id differences

How we have tested the changes on our side:

- Integration tests have been run against clean installations of Magento
2.4.4, 2.4.5 and 2.4.6 and they all seem to pass
- All Magento Functional Testing Framework tests listed here have been
run against clean installations of Magento 2.4.4, 2.4.5 and 2.4.6 and
they all seem to pass:
https://developer.adobe.com/commerce/marketplace/guides/sellers/mftf-magento/
- One of our internal QA engineers has checked module functionality and
connections to Custobar API in Magento 2.4.6 environment and no issues
were found